re-organize discovery tests to partition based on network
